Added feature that allowed creation of users
Still in the process of allowing users to be created. Was successful in modifying the method but will need to modify the method for the parent method. Also this will require authenticating the login using the database as well. (Still have yet to do that... should be next step and then polish the user creation and login)